refactored useCreateChartTitle hook
