+++ frontend/src/cards/DisparityBarChartCard.tsx
@@ -23,7 +23,7 @@ import { CAWP_DETERMINANTS } from "../data/variables/CawpProvider";
 import { useGuessPreloadHeight } from "../utils/hooks/useGuessPreloadHeight";
 import { reportProviderSteps } from "../reports/ReportProviderSteps";
 import { ScrollableHashId } from "../utils/hooks/useStepObserver";
-import { useCreateChartTitle } from "../utils/hooks/useCreateTitle";
+import { useCreateChartTitle } from "../utils/hooks/useCreateChartTitle";
 
 export interface DisparityBarChartCardProps {
   key?: string;
