+++ frontend/src/cards/MapCard.tsx
@@ -50,7 +50,7 @@ import { createSubtitle } from "../charts/utils";
 import { useLocation } from "react-router-dom";
 import { reportProviderSteps } from "../reports/ReportProviderSteps";
 import { ScrollableHashId } from "../utils/hooks/useStepObserver";
-import { useCreateChartTitle } from "../utils/hooks/useCreateTitle";
+import { useCreateChartTitle } from "../utils/hooks/useCreateChartTitle";
 
 const SIZE_OF_HIGHEST_LOWEST_RATES_LIST = 5;
